backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0

sqlalchemy>=2.0.0
aiosqlite>=0.19.0

anthropic>=0.40.0

yfinance>=0.2.40
finnhub-python>=2.4.20
fredapi>=0.5.2
pandas>=2.2.0
numpy>=1.26.0

apscheduler>=3.10.0

tenacity>=8.2.0
orjson>=3.9.0
